+++ app/src/main/res/xml/network_security_config.xml
@@ -2,6 +2,6 @@
 <network-security-config>
     <domain-config cleartextTrafficPermitted="true">
         <domain includeSubdomains="true">15.164.105.26</domain>
-        <domain includeSubdomains="true">good-place</domain>
+        <domain includeSubdomains="true">good-place.shop</domain>
     </domain-config>
 </network-security-config>
